+++ src/detectmatelibrary/common/_core_op/_fit_logic.py
